refactor: extract request phases into dedicated modules
Move phase-specific logic out of lib.rs into phases/ modules:
- phases/on_request.rs — phase 1 (headers) and phase 2 (body) interceptors
- phases/before_upstream.rs — header adjustments and before_upstream interceptors
- phases/on_response.rs — on_response and on_response_body interceptors
- phases/upstream_peer.rs — peer resolution, cancellation check, timeout setup

lib.rs is now a thin orchestration layer delegating to phase modules.
