+++ plenum-core/src/lib.rs
@@ -8,13 +8,11 @@ use bytes::{BufMut, Bytes, BytesMut};
 use config::Config;
 use gateway_error::GatewayError;
 use http::Method;
-use interceptor::{InterceptorOutput, request_input_from_parts, response_input_from_parts};
 use path_match::{HookHandle, OperationSchemas, RouteEntry, Upstream, build_router};
 use pingora_core::upstreams::peer::HttpPeer;
-use pingora_http::{RequestHeader, ResponseHeader};
+use pingora_http::ResponseHeader;
 use pingora_proxy::{FailToProxy, ProxyHttp, Session};
 use tokio_util::sync::CancellationToken;
-use tracing::Instrument;
 
 pub mod config;
 mod ctx;
@@ -23,17 +21,13 @@ mod headers;
 pub mod interceptor;
 mod openapi;
 pub mod path_match;
+mod phases;
 mod proxy_utils;
 pub mod request_timeout;
 pub mod upstream_http;
 mod upstream_plugin;
 
 pub use ctx::GatewayCtx;
-use headers::apply_header_modifications;
-use proxy_utils::{
-    call_interceptor, call_interceptor_blocking, js_body_from_content_type, js_body_to_bytes,
-    merge_options,
-};
 
 pub struct Plenum {
     router: path_match::PlenumRouter,
@@ -42,7 +36,11 @@ pub struct Plenum {
 /// Compute effective interceptor timeout from remaining request budget.
 /// Cancels the request if budget is exhausted. Falls back to the per-interceptor
 /// timeout when no overall request timeout is active.
-fn effective_timeout(ctx: &GatewayCtx, op: &OperationSchemas, hook: &HookHandle) -> Duration {
+pub(crate) fn effective_timeout(
+    ctx: &GatewayCtx,
+    op: &OperationSchemas,
+    hook: &HookHandle,
+) -> Duration {
     let Some(start) = ctx.request_start else {
         return hook.timeout;
     };
@@ -53,105 +51,6 @@ fn effective_timeout(ctx: &GatewayCtx, op: &OperationSchemas, hook: &HookHandle)
             Duration::ZERO
         }
     }
-}
-
-/// Look up the matched operation from context, borrowing only the route and method fields.
-/// Passing the fields explicitly (rather than `&GatewayCtx`) lets other fields remain
-/// independently mutable while the returned reference is alive.
-fn matched_op<'a>(
-    matched_route: &'a Option<Arc<RouteEntry>>,
-    matched_method: &Option<Method>,
-) -> Option<&'a OperationSchemas> {
-    let route = matched_route.as_ref()?;
-    let method = matched_method.as_ref()?;
-    route.operations.get(method)
-}
-
-/// Run on_request phase 1 interceptors (headers only, null body).
-/// Returns `Ok(true)` if the request was short-circuited (respond or error).
-///
-/// When `budget_cap` is true, each interceptor call is capped to the remaining
-/// request budget and the cancellation token is checked before each call.
-/// When false (plugin routes inside a timeout wrapper), interceptors use their
-/// own per-interceptor timeout and no budget check is performed.
-async fn run_on_request_phase1(
-    session: &mut Session,
-    ctx: &GatewayCtx,
-    op: &OperationSchemas,
-    budget_cap: bool,
-) -> pingora_core::Result<bool> {
-    for hook in &op.interceptors.on_request {
-        let timeout = if budget_cap {
-            let t = effective_timeout(ctx, op, hook);
-            if ctx.cancellation.is_cancelled() {
-                session
-                    .respond_error_with_body(
-                        504,
-                        GatewayError::gateway_timeout("request timeout exceeded").body(),
-                    )
-                    .await
-                    .ok();
-                return Ok(true);
-            }
-            t
-        } else {
-            hook.timeout
-        };
-
-        let input = request_input_from_parts(
-            &session.req_header().method,
-            &session.req_header().uri,
-            &session.req_header().headers,
-            ctx.path_params.clone(),
-            op.operation_meta.clone(),
-        );
-        let mut input_json = serde_json::to_value(&input).unwrap();
-        merge_options(&mut input_json, hook.options.as_ref());
-
-        let span = tracing::debug_span!(
-            "interceptor_call",
-            hook = "on_request",
-            function = hook.function.as_str()
-        );
-        match call_interceptor(
-            hook.runtime.as_ref(),
-            &hook.function,
-            input_json,
-            None,
-            timeout,
-        )
-        .instrument(span)
-        .await
-        {
-            Ok((InterceptorOutput::Continue { headers, .. }, _)) => {
-                if let Some(mods) = headers {
-                    apply_header_modifications(session.req_header_mut(), &mods);
-                }
-            }
-            Ok((InterceptorOutput::Respond { status, .. }, body_out)) => {
-                session
-                    .respond_error_with_body(
-                        status,
-                        body_out.map(js_body_to_bytes).unwrap_or_default(),
-                    )
-                    .await
-                    .ok();
-                return Ok(true);
-            }
-            Err(e) => {
-                log::error!("on_request interceptor error: {}", e);
-                session
-                    .respond_error_with_body(
-                        500,
-                        GatewayError::internal(format!("interceptor error: {}", e)).body(),
-                    )
-                    .await
-                    .ok();
-                return Ok(true);
-            }
-        }
-    }
-    Ok(false)
 }
 
 #[async_trait]
@@ -201,9 +100,6 @@ impl ProxyHttp for Plenum {
             .map(|(k, v)| (k.to_string(), v.to_string()))
             .collect();
 
-        // Clone the Arc (cheap) so that `op` and `plugin` borrow from a local variable rather
-        // than from ctx. This lets us later pass ctx mutably to upstream_plugin::dispatch while
-        // op and plugin are still in scope.
         let Some(route_arc) = ctx.matched_route.clone() else {
             return Ok(false);
         };
@@ -216,12 +112,11 @@ impl ProxyHttp for Plenum {
 
         ctx.request_start = Some(Instant::now());
 
-        // For plugin routes: wrap on_request phase 1 + dispatch in a single timeout.
-        // When the timeout fires the future is dropped, cancelling all in-flight work.
+        // Plugin routes: wrap on_request phase 1 + dispatch in a single timeout.
         if let Upstream::Plugin(plugin) = &route_arc.upstream {
             let backend_config = op.backend_config.clone();
             return match pingora_timeout::timeout(op.request_timeout, async {
-                if run_on_request_phase1(session, ctx, op, false).await? {
+                if phases::on_request::run_phase1(session, ctx, op, false).await? {
                     return Ok(true);
                 }
                 upstream_plugin::dispatch(session, ctx, op, plugin, backend_config).await
@@ -244,12 +139,12 @@ impl ProxyHttp for Plenum {
             };
         }
 
-        // Phase 1 of on_request for HTTP/Static routes: budget-capped interceptor calls.
-        if run_on_request_phase1(session, ctx, op, true).await? {
+        // HTTP/Static routes: budget-capped on_request phase 1.
+        if phases::on_request::run_phase1(session, ctx, op, true).await? {
             return Ok(true);
         }
 
-        // For static routes: write the pre-built response and short-circuit.
+        // Static routes: write the pre-built response and short-circuit.
         if let Upstream::Static(static_resp) = &route_arc.upstream {
             let mut resp_header = pingora_http::ResponseHeader::build(static_resp.status, None)
                 .map_err(|e| {
@@ -300,17 +195,23 @@ impl ProxyHttp for Plenum {
     where
         Self::CTX: Send + Sync,
     {
-        let Some(op) = matched_op(&ctx.matched_route, &ctx.matched_method) else {
+        // Clone the Arc so `op` doesn't borrow from `ctx`, allowing ctx to be
+        // passed mutably to the phase handler.
+        let Some(route_arc) = ctx.matched_route.clone() else {
+            return Ok(());
+        };
+        let Some(method) = ctx.matched_method.clone() else {
+            return Ok(());
+        };
+        let Some(op) = route_arc.operations.get(&method) else {
             return Ok(());
         };
 
-        // Skip if no on_request interceptors are configured.
-        // Plugin routes are fully handled in request_filter and never reach here.
         if op.interceptors.on_request.is_empty() {
             return Ok(());
         }
 
-        // Buffer chunks, suppressing forwarding until we're done processing
+        // Buffer chunks until end_of_stream
         if let Some(b) = body {
             ctx.request_body_buf.put(b.as_ref());
             b.clear();
@@ -319,92 +220,12 @@ impl ProxyHttp for Plenum {
         if end_of_stream {
             let buf = ctx.request_body_buf.split().freeze();
 
-            // Step 2: Phase 2 of on_request (body access). Only fires for non-empty bodies;
-            // phase 1 in request_filter already ran. Here we apply body changes only --
-            // header modifications are intentionally ignored since upstream headers were
-            // already built before request_body_filter runs.
-            let final_buf = if !op.interceptors.on_request.is_empty() && !buf.is_empty() {
-                let content_type = session
-                    .req_header()
-                    .headers
-                    .get(http::header::CONTENT_TYPE)
-                    .and_then(|v| v.to_str().ok())
-                    .map(|s| s.to_string());
-                let mut current_buf = buf;
-                for hook in &op.interceptors.on_request {
-                    let timeout = effective_timeout(ctx, op, hook);
-                    if ctx.cancellation.is_cancelled() {
-                        session
-                            .respond_error_with_body(
-                                504,
-                                GatewayError::gateway_timeout("request timeout exceeded").body(),
-                            )
-                            .await
-                            .ok();
-                        ctx.filter_responded = true;
-                        return Ok(());
-                    }
-
-                    let js_body = js_body_from_content_type(content_type.as_deref(), &current_buf);
-                    let input = request_input_from_parts(
-                        &session.req_header().method,
-                        &session.req_header().uri,
-                        &session.req_header().headers,
-                        ctx.path_params.clone(),
-                        op.operation_meta.clone(),
-                    );
-                    let mut input_json = serde_json::to_value(&input).unwrap();
-                    merge_options(&mut input_json, hook.options.as_ref());
-                    let span = tracing::debug_span!(
-                        "interceptor_call",
-                        hook = "on_request_body",
-                        function = hook.function.as_str()
-                    );
-                    match call_interceptor(
-                        hook.runtime.as_ref(),
-                        &hook.function,
-                        input_json,
-                        js_body,
-                        timeout,
-                    )
-                    .instrument(span)
-                    .await
-                    {
-                        Ok((InterceptorOutput::Continue { .. }, body_out)) => {
-                            current_buf = body_out.map(js_body_to_bytes).unwrap_or(current_buf);
-                        }
-                        Ok((InterceptorOutput::Respond { status, .. }, body_out)) => {
-                            session
-                                .respond_error_with_body(
-                                    status,
-                                    body_out.map(js_body_to_bytes).unwrap_or_default(),
-                                )
-                                .await
-                                .ok();
-                            ctx.filter_responded = true;
-                            return Ok(());
-                        }
-                        Err(e) => {
-                            log::error!("on_request interceptor error: {}", e);
-                            session
-                                .respond_error_with_body(
-                                    500,
-                                    GatewayError::internal(format!("interceptor error: {}", e))
-                                        .body(),
-                                )
-                                .await
-                                .ok();
-                            ctx.filter_responded = true;
-                            return Ok(());
-                        }
-                    }
-                }
-                current_buf
-            } else {
-                buf
+            let final_buf = match phases::on_request::run_phase2_body(session, ctx, op, buf).await?
+            {
+                Some(b) => b,
+                None => return Ok(()), // short-circuited (filter_responded already set)
             };
 
-            // Step 3: Restore body for upstream forwarding (only if non-empty)
             if !final_buf.is_empty() {
                 *body = Some(final_buf);
             }
@@ -416,7 +237,7 @@ impl ProxyHttp for Plenum {
     async fn upstream_request_filter(
         &self,
         _session: &mut Session,
-        upstream_request: &mut RequestHeader,
+        upstream_request: &mut pingora_http::RequestHeader,
         ctx: &mut Self::CTX,
     ) -> pingora_core::Result<()>
     where
@@ -425,79 +246,7 @@ impl ProxyHttp for Plenum {
         let Some(op) = matched_op(&ctx.matched_route, &ctx.matched_method) else {
             return Ok(());
         };
-
-        // When on_request is configured, the body may be modified in request_body_filter.
-        // Replace Content-Length with Transfer-Encoding: chunked so pingora will stream
-        // the (possibly resized) body to the upstream without a fixed length constraint.
-        // Only apply to requests that actually carry a body (have Content-Length set).
-        if !op.interceptors.on_request.is_empty()
-            && upstream_request
-                .headers
-                .contains_key(http::header::CONTENT_LENGTH)
-        {
-            upstream_request.remove_header(&http::header::CONTENT_LENGTH);
-            upstream_request
-                .insert_header(http::header::TRANSFER_ENCODING, "chunked")
-                .ok();
-        }
-
-        // When on_response_body is configured, we need to buffer and inspect the response body.
-        // Prevent gzip encoding from the upstream so we receive raw bytes.
-        if !op.interceptors.on_response_body.is_empty() {
-            upstream_request
-                .insert_header("accept-encoding", "identity")
-                .ok();
-        }
-
-        for hook in &op.interceptors.before_upstream {
-            let timeout = effective_timeout(ctx, op, hook);
-            if ctx.cancellation.is_cancelled() {
-                log::warn!("request timeout exceeded during before_upstream interceptors");
-                break;
-            }
-
-            let input = request_input_from_parts(
-                &upstream_request.method,
-                &upstream_request.uri,
-                &upstream_request.headers,
-                ctx.path_params.clone(),
-                op.operation_meta.clone(),
-            );
-            let mut input_json = serde_json::to_value(&input).unwrap();
-            merge_options(&mut input_json, hook.options.as_ref());
-
-            let span = tracing::debug_span!(
-                "interceptor_call",
-                hook = "before_upstream",
-                function = hook.function.as_str()
-            );
-            match call_interceptor(
-                hook.runtime.as_ref(),
-                &hook.function,
-                input_json,
-                None,
-                timeout,
-            )
-            .instrument(span)
-            .await
-            {
-                Ok((InterceptorOutput::Continue { headers, .. }, _)) => {
-                    if let Some(mods) = &headers {
-                        apply_header_modifications(upstream_request, mods);
-                    }
-                }
-                Ok((InterceptorOutput::Respond { .. }, _)) => {
-                    log::warn!(
-                        "before_upstream interceptor returned 'respond' -- ignoring (request already committed to upstream)"
-                    );
-                }
-                Err(e) => {
-                    log::error!("before_upstream interceptor error: {}", e);
-                }
-            }
-        }
-
-        Ok(())
+        phases::before_upstream::run(upstream_request, ctx, op).await
     }
 
     async fn response_filter(
@@ -513,56 +262,7 @@ impl ProxyHttp for Plenum {
             return Ok(());
         };
 
-        for hook in &op.interceptors.on_response {
-            let timeout = effective_timeout(ctx, op, hook);
-            if ctx.cancellation.is_cancelled() {
-                log::warn!("request timeout exceeded during on_response interceptors");
-                break;
-            }
-
-            let input = response_input_from_parts(
-                upstream_response.status,
-                &upstream_response.headers,
-                op.operation_meta.clone(),
-            );
-            let mut input_json = serde_json::to_value(&input).unwrap();
-            merge_options(&mut input_json, hook.options.as_ref());
-
-            let span = tracing::debug_span!(
-                "interceptor_call",
-                hook = "on_response",
-                function = hook.function.as_str()
-            );
-            match call_interceptor(
-                hook.runtime.as_ref(),
-                &hook.function,
-                input_json,
-                None,
-                timeout,
-            )
-            .instrument(span)
-            .await
-            {
-                Ok((InterceptorOutput::Continue { status, headers }, _)) => {
-                    if let Some(code) = status
-                        && let Ok(status_code) = http::StatusCode::from_u16(code)
-                    {
-                        upstream_response.set_status(status_code).ok();
-                    }
-                    if let Some(mods) = &headers {
-                        apply_header_modifications(upstream_response, mods);
-                    }
-                }
-                Ok((InterceptorOutput::Respond { .. }, _)) => {
-                    log::warn!(
-                        "on_response interceptor returned 'respond' -- ignoring (response already in flight)"
-                    );
-                }
-                Err(e) => {
-                    log::error!("on_response interceptor error: {}", e);
-                }
-            }
-        }
+        phases::on_response::run(upstream_response, ctx, op).await?;
 
         // When on_response_body is configured, strip Content-Length (body size may change)
         // and store metadata in ctx for use in upstream_response_body_filter.
@@ -597,7 +297,7 @@ impl ProxyHttp for Plenum {
             return Ok(None);
         }
 
-        // Buffer chunks, suppressing forwarding until end_of_stream
+        // Buffer chunks until end_of_stream
         if let Some(b) = body {
             ctx.response_body_buf.put(b.as_ref());
             b.clear();
@@ -607,55 +307,8 @@ impl ProxyHttp for Plenum {
             let buf = ctx.response_body_buf.split().freeze();
             let status = ctx.upstream_response_status.unwrap_or(http::StatusCode::OK);
 
-            let final_buf = tokio::task::block_in_place(|| {
-                let mut current_buf = buf;
-                for hook in &op.interceptors.on_response_body {
-                    let timeout = effective_timeout(ctx, op, hook);
-                    if ctx.cancellation.is_cancelled() {
-                        log::warn!("request timeout exceeded during on_response_body interceptors");
-                        break;
-                    }
-
-                    let js_body = js_body_from_content_type(
-                        ctx.upstream_response_content_type.as_deref(),
-                        &current_buf,
-                    );
-                    let input = response_input_from_parts(
-                        status,
-                        &http::HeaderMap::new(),
-                        op.operation_meta.clone(),
-                    );
-                    let mut input_json = serde_json::to_value(&input).unwrap();
-                    merge_options(&mut input_json, hook.options.as_ref());
-
-                    let _span = tracing::debug_span!(
-                        "interceptor_call",
-                        hook = "on_response_body",
-                        function = hook.function.as_str()
-                    )
-                    .entered();
-                    match call_interceptor_blocking(
-                        hook.runtime.as_ref(),
-                        &hook.function,
-                        input_json,
-                        js_body,
-                        timeout,
-                    ) {
-                        Ok((InterceptorOutput::Continue { .. }, body_out)) => {
-                            current_buf = body_out.map(js_body_to_bytes).unwrap_or(current_buf);
-                        }
-                        Ok((InterceptorOutput::Respond { .. }, _)) => {
-                            log::warn!(
-                                "on_response_body interceptor returned 'respond' -- ignoring"
-                            );
-                        }
-                        Err(e) => {
-                            log::error!("on_response_body interceptor error: {}", e);
-                        }
-                    }
-                }
-                current_buf
-            });
+            let final_buf =
+                tokio::task::block_in_place(|| phases::on_response::run_body(ctx, op, buf, status));
 
             *body = Some(final_buf);
         }
@@ -668,44 +321,7 @@ impl ProxyHttp for Plenum {
         _session: &mut Session,
         ctx: &mut Self::CTX,
     ) -> pingora_core::Result<Box<HttpPeer>> {
-        if ctx.filter_responded {
-            return Err(pingora_core::Error::new(
-                pingora_core::ErrorType::HTTPStatus(400),
-            ));
-        }
-        if ctx.cancellation.is_cancelled() {
-            return Err(pingora_core::Error::explain(
-                pingora_core::ErrorType::ConnectTimedout,
-                "request cancelled",
-            ));
-        }
-        let route = ctx
-            .matched_route
-            .as_ref()
-            .ok_or_else(|| pingora_core::Error::new(pingora_core::ErrorType::InternalError))?;
-        match &route.upstream {
-            crate::path_match::Upstream::Http(peer) => {
-                let mut peer = Box::new(*peer.clone());
-
-                // Set upstream timeouts to remaining request budget so pingora races
-                // the upstream I/O against the overall request timeout.
-                if let (Some(start), Some(op)) = (
-                    ctx.request_start,
-                    matched_op(&ctx.matched_route, &ctx.matched_method),
-                ) {
-                    request_timeout::apply_to_peer(&mut peer, start, op.request_timeout)?;
-                }
-
-                Ok(peer)
-            }
-            crate::path_match::Upstream::Plugin(_) | crate::path_match::Upstream::Static(_) => {
-                // Should never be reached -- plugin and static routes return Ok(true) from
-                // request_filter, skipping upstream_peer entirely. This branch is a safety net.
-                Err(pingora_core::Error::new(
-                    pingora_core::ErrorType::InternalError,
-                ))
-            }
-        }
+        phases::upstream_peer::resolve(ctx)
     }
 
     async fn fail_to_proxy(
@@ -717,8 +333,6 @@ impl ProxyHttp for Plenum {
     where
         Self::CTX: Send + Sync,
     {
-        // When an overall request timeout is active and the upstream error is a timeout,
-        // return 504 with a JSON body instead of pingora's default error page.
         if ctx.request_start.is_some() && request_timeout::is_timeout_error(e) {
             session
                 .respond_error_with_body(
@@ -733,13 +347,22 @@ impl ProxyHttp for Plenum {
             };
         }
 
-        // Fall through to default pingora behaviour for non-timeout errors.
         session.respond_error(502).await.ok();
         FailToProxy {
             error_code: 502,
             can_reuse_downstream: false,
         }
     }
+}
+
+/// Look up the matched operation from context.
+fn matched_op<'a>(
+    matched_route: &'a Option<Arc<RouteEntry>>,
+    matched_method: &Option<Method>,
+) -> Option<&'a OperationSchemas> {
+    let route = matched_route.as_ref()?;
+    let method = matched_method.as_ref()?;
+    route.operations.get(method)
 }
 
 pub fn build_gateway(config: &Config, config_path: &str) -> Result<Plenum, Box<dyn Error>> {
